medical-guideline-validation-main/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# OpenAI
openai>=1.30.0
httpx>=0.27.0
httpcore>=1.0.0
# Environment Variables
python-dotenv==1.0.0

# RAG and Vector Database
langchain==0.1.6
langchain-community==0.0.20
chromadb==0.4.22

# Embeddings and NLP
sentence-transformers==2.3.1
transformers==4.37.2

# PDF Processing
pypdf==4.0.1
PyPDF2==3.0.1

# HTTP Requests
requests==2.31.0

# Utilities
python-multipart==0.0.6